Backend_Code/djangowebapp/requirements.txt
mysqlclient==2.2.0
Django==4.1
djangorestframework==3.13.1
sqlparse==0.4.2

